requirements.txt
pylint
pytest
requests
psycopg2
python-dotenv
flask
beautifulsoup4
altair
streamlit